requirements.txt
# Core dependencies for event management system

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Database
psycopg2-binary>=2.9.0

# Environment variables
python-dotenv>=1.0.0

# Jupyter notebook
jupyter>=1.0.0
notebook>=7.0.0

# Visualization (for event_analysis.py)
matplotlib>=3.7.0
seaborn>=0.12.0
